+++ packages/user/Fractals/service/src/tables/mod.rs
@@ -117,7 +117,7 @@ pub mod tables {
         pub member: AccountNumber,
         #[graphql(skip)]
         pub score: u32,
-        pub pending_score: Option<u32>,
+        pub pending_score: Option<u8>,
         pub created_at: psibase::TimePointSec,
     }
